+++ xuance/torch/learners/policy_gradient/sac_learner.py
@@ -41,7 +41,7 @@ class SAC_Learner(Learner):
         act_batch = torch.as_tensor(samples['actions'], device=self.device)
         next_batch = torch.as_tensor(samples['obs_next'], device=self.device)
         rew_batch = torch.as_tensor(samples['rewards'], device=self.device)
-        ter_batch = torch.as_tensor(samples['terminals'], device=self.device)
+        ter_batch = torch.as_tensor(samples['terminals'], dtype=torch.float, device=self.device)
 
         # actor update
         log_pi, policy_q_1, policy_q_2 = self.policy.Qpolicy(obs_batch)
